backend/requirements.txt
django>=6.0
djangorestframework>=3.16
djangorestframework-simplejwt>=5.5
django-cors-headers>=4.0
Pillow>=10.0
python-decouple>=3.8
psycopg2-binary>=2.9
django-filter>=24.2
